+++ public/src/pages/live.vue
@@ -1,23 +1,27 @@
 <template>
     <layout>
         <div class="title">
-            <h1>Live: {{data.name}}</h1>
+            <h1>Live: {{ data.name }}</h1>
             <div class="menu">
-                <button v-on:click="loadLive"><i class="fas fa-sync"></i> {{$t("refresh")}}</button>
-                <button class="primary" v-on:click="$router.push('/')"><i class="fas fa-arrow-left"></i> {{$t("back")}}</button>
+                <select v-model="refreshRate">
+                    <option v-for="o in [2, 5, 10, 15, 20]" v-bind:value="o">{{ o }} sec</option>
+                </select>
+                <button v-on:click="loadLive"><i class="fas fa-sync"></i> {{ $t("refresh") }}</button>
+                <button class="primary" v-on:click="$router.push('/')"><i class="fas fa-arrow-left"></i>
+                    {{ $t("back") }}</button>
             </div>
         </div>
 
         <div class="content">
             <div class="header">
-                <div id="state"><strong>Status:</strong> {{data.live.serverState}}</div>
-                <div id="track"><strong>Track:</strong> {{data.live.track}}</div>
+                <div id="state"><strong>Status:</strong> {{ data.live.serverState }}</div>
+                <div id="track"><strong>Track:</strong> {{ data.live.track }}</div>
                 <div id="phase">
-                    <strong>Phase:</strong> {{data.live.sessionType}} ({{data.live.sessionPhase}})
-                    <span v-if="data.live.sessionRemaining > 0">[{{data.live.sessionRemaining}} min]</span>
+                    <strong>Phase:</strong> {{ data.live.sessionType }} ({{ data.live.sessionPhase }})
+                    <span v-if="data.live.sessionRemaining > 0">[{{ data.live.sessionRemaining }} min]</span>
                 </div>
-                <div id="nrdrivers"><strong>Drivers:</strong> {{data.live.nrClients}}</div>
-                <div id="updatedat"><strong>Last Update:</strong> {{new Date(data.live.updatedAt).toLocaleString()}}</div>
+                <div id="nrdrivers"><strong>Drivers:</strong> {{ data.live.nrClients }}</div>
+                <div id="updatedat"><strong>Last Update:</strong> {{ new Date(data.live.updatedAt).toLocaleString() }}</div>
             </div>
 
             <div class="body">
@@ -38,23 +42,21 @@
                         <th>Flags</th>
                     </tr>
 
-                    <tr v-for="(car, carId) in orderedCars" :key="carId"
-                        v-on:click="setShowLaps(car.carID)"
-                        v-bind:class="{'tbl-row': true, active: car.carID === showLaps}"
-                    >
-                        <td>{{carId+1}}</td>
-                        <td>{{car.currentDriver ? car.currentDriver.name.toUpperCase() : car.carID}}</td>
-                        <td>{{car.raceNumber}}</td>
-                        <td>{{car.carModel}}</td>
-                        <td>{{car.nrLaps}}</td>
-                        <td>{{car.fuel}}</td>
-                        <td>{{msToTime(car.bestLapMS)}}</td>
-                        <td v-bind:class="{invalid: car.currLap.flags > 0}">{{msToTime(car.lastLapMS)}}</td>
-                        <td v-bind:class="{invalid: car.currLap.flags > 0}">{{car.currLap.s1}}</td>
-                        <td v-bind:class="{invalid: car.currLap.flags > 0}">{{car.currLap.s2}}</td>
-                        <td v-bind:class="{invalid: car.currLap.flags > 0}">{{car.currLap.s3}}</td>
-                        <td v-if="data.live.sessionType == 'Race'">{{calcGap(carId)}}</td>
-                        <td v-bind:class="{invalid: car.currLap.flags > 0}">
+                    <tr v-for="(car, carId) in orderedCars" :key="carId" v-on:click="setShowLaps(car.carID)"
+                        v-bind:class="{ 'tbl-row': true, active: car.carID === showLaps }">
+                        <td>{{ carId + 1 }}</td>
+                        <td>{{ car.currentDriver ? car.currentDriver.name.toUpperCase() : car.carID }}</td>
+                        <td>{{ car.raceNumber }}</td>
+                        <td>{{ car.carModel }}</td>
+                        <td>{{ car.nrLaps }}</td>
+                        <td>{{ car.fuel }}</td>
+                        <td>{{ msToTime(car.bestLapMS) }}</td>
+                        <td v-bind:class="{ invalid: car.currLap.flags > 0 }">{{ msToTime(car.lastLapMS) }}</td>
+                        <td v-bind:class="{ invalid: car.currLap.flags > 0 }">{{ car.currLap.s1 }}</td>
+                        <td v-bind:class="{ invalid: car.currLap.flags > 0 }">{{ car.currLap.s2 }}</td>
+                        <td v-bind:class="{ invalid: car.currLap.flags > 0 }">{{ car.currLap.s3 }}</td>
+                        <td v-if="data.live.sessionType == 'Race'">{{ calcGap(carId) }}</td>
+                        <td v-bind:class="{ invalid: car.currLap.flags > 0 }">
                             <i class="fas fa-cut" v-if="car.currLap.hasCut" title="Has Cut"></i>
                             <i class="fas fa-sign-in-alt" v-if="car.currLap.inLap" title="In Lap"></i>
                             <i class="fas fa-sign-out-alt" v-if="car.currLap.outLap" title="Out Lap"></i>
@@ -64,7 +66,7 @@
                 </table>
 
                 <div id="laps" v-if="showLaps">
-                    <h3>Car {{showLapsCar.raceNumber}} Laps</h3>
+                    <h3>Car {{ showLapsCar.raceNumber }} Laps</h3>
 
                     <table>
                         <tr class="tbl-header">
@@ -79,15 +81,18 @@
                             <th>Flags</th>
                         </tr>
 
-                        <tr v-for="(lap, i) in showLapsCar.laps" :key="i" v-bind:class="{ 'tbl-row': true, invalid: lap.flags > 0, best: lap.lapTimeMS === showLapsCar.bestLapMS }">
-                            <td>{{i+1}}</td>
-                            <td>{{showLapsCar.drivers[lap.driverIndex] ? showLapsCar.drivers[lap.driverIndex].name : '--'}}</td>
-                            <td>{{lap.fuel}}</td>
-                            <td>{{msToTime(lap.lapTimeMS)}}</td>
-                            <td>{{lap.s1}}</td>
-                            <td>{{lap.s2}}</td>
-                            <td>{{lap.s3}}</td>
-                            <td align="right">{{calcDelta(i)}}</td>
+                        <tr v-for="(lap, i) in showLapsCar.laps" :key="i"
+                            v-bind:class="{ 'tbl-row': true, invalid: lap.flags > 0, best: lap.lapTimeMS === showLapsCar.bestLapMS }">
+                            <td>{{ i + 1 }}</td>
+                            <td>{{ showLapsCar.drivers[lap.driverIndex] ? showLapsCar.drivers[lap.driverIndex].name : '--'
+                            }}
+                            </td>
+                            <td>{{ lap.fuel }}</td>
+                            <td>{{ msToTime(lap.lapTimeMS) }}</td>
+                            <td>{{ lap.s1 }}</td>
+                            <td>{{ lap.s2 }}</td>
+                            <td>{{ lap.s3 }}</td>
+                            <td align="right">{{ calcDelta(i) }}</td>
                             <td>
                                 <i class="fas fa-cut" v-if="lap.hasCut" title="Has Cut"></i>
                                 <i class="fas fa-sign-in-alt" v-if="lap.inLap" title="In Lap"></i>
@@ -99,13 +104,41 @@
                 </div>
 
                 <div id="chat">
-                    <h3>Chat</h3>
+                    <h3>Events</h3>
 
-                    <div class="message" v-for="item in data.live.chats.slice().reverse()" :key="item.ts">
-                        <div class="ts">{{new Date(item.ts).toLocaleString()}}</div>
-                        <div class="name">{{item.name}}:</div> 
-                        <div class="msg">{{item.message}}</div>
+                    <div class=" box server-settings-container four-columns">
+                        <div>
+                            <checkbox label="Hide damage report" v-model="hideDamage" />
+                        </div>
+                        <div>
+                            <checkbox label="Hide session changes" v-model="hideSession" />
+                        </div>
+                        <div>
+                            <selection label="Show last events" :options="eventsOptions" v-model="nrEvents" />
+                        </div>
                     </div>
+
+                    <div v-for="item in eventHistories" :key="item.id">
+                        <div class="message" v-if="item.type == 'chat'">
+                            <div class="ts">{{ timeSince(new Date(item.ts)) }}</div>
+                            <div class="name">{{ item.data.name }}:</div>
+                            <div class="msg">{{ item.data.message }}</div>
+                        </div>
+
+                        <div class="message session" v-if="item.type == 'session'">
+                            <div class="ts">{{ timeSince(new Date(item.ts)) }}</div>
+                            <div class="name">Session:</div>
+                            <div class="msg">{{ item.data.sessionType }} - {{ item.data.sessionPhase }} ({{
+                                item.data.sessionRemaining }} min)</div>
+                        </div>
+
+                        <div class="message damage" v-if="item.type == 'damage'">
+                            <div class="ts">{{ timeSince(new Date(item.ts)) }}</div>
+                            <div class="name">{{ item.data.name.toUpperCase() }}:</div>
+                            <div class="msg">Damage Report #{{ item.data.raceNumber }}</div>
+                        </div>
+                    </div>
+
                 </div>
             </div>
         </div>
@@ -114,19 +147,31 @@
 
 <script>
 import axios from "axios";
-import {layout} from "../components";
+import { layout, checkbox, selection } from "../components";
 import _ from "lodash";
+import moment from "moment";
 
 let toId = null;
 
 export default {
     name: "live",
-    components: {layout},
+    components: { layout, checkbox, selection },
     data() {
         return {
             id: "",
             showLaps: null,
             showLapsCar: null,
+            refreshRate: 10,
+            hideDamage: false,
+            hideSession: false,
+            nrEvents: 100,
+            eventsOptions: [
+                { label: "30", value: 30 },
+                { label: "50", value: 50 },
+                { label: "100", value: 100 },
+                { label: "150", value: 150 },
+                { label: "200", value: 200 }
+            ],
             data: {
                 name: "",
                 track: "",
@@ -137,7 +182,7 @@ export default {
                     sessionPhase: "",
                     sessionRemaining: 0,
                     cars: {},
-                    chats: []
+                    history: []
                 }
             },
         };
@@ -157,11 +202,28 @@ export default {
             const ordered = _.orderBy(this.data.live.cars, "position");
             return _.filter(ordered, (o) => { return o.currentDriver !== null });
         },
-        classObject: function(lap) {
+        classObject: function (lap) {
             return {
                 "tbl-row": true,
                 "lap-invalid": lap.flags > 0
             }
+        },
+        eventHistories: function () {
+            const hd = this.hideDamage;
+            const hs = this.hideSession;
+            const data = _.filter(this.data.live.history, function (o) {
+                if (hd && o.type == "damage") {
+                    return false
+                }
+
+                if (hs && o.type == "session") {
+                    return false
+                }
+
+                return true;
+            })
+
+            return _.slice(_.reverse(data), 0, this.nrEvents);
         }
     },
     methods: {
@@ -186,7 +248,7 @@ export default {
             this.loadLive();
             toId = setTimeout(() => {
                 this.refreshList();
-            }, 10000);
+            }, this.refreshRate * 1000);
         },
         setShowLaps(carID, refresh) {
             if (refresh === undefined && this.showLaps === carID) {
@@ -203,7 +265,7 @@ export default {
                 return {};
             }
 
-            return laps[laps.length-1]
+            return laps[laps.length - 1]
         },
         msToTime(ms) {
             if (ms === 0 || ms === undefined) {
@@ -220,7 +282,7 @@ export default {
             const s = ms / 1000;
             const m = s / 60;
 
-            return `${sign}${Math.floor(m % 60)}:${_.padStart(Math.floor(s%60).toString(), 2, '0')}.${_.padStart(Math.floor(ms%1000).toString(), 3, '0')}`;
+            return `${sign}${Math.floor(m % 60)}:${_.padStart(Math.floor(s % 60).toString(), 2, '0')}.${_.padStart(Math.floor(ms % 1000).toString(), 3, '0')}`;
         },
         calcDelta(idx) {
             const bestLap = this.showLapsCar.bestLapMS;
@@ -234,7 +296,7 @@ export default {
             }
 
             const curr = this.orderedCars[idx];
-            const prev = this.orderedCars[idx-1];
+            const prev = this.orderedCars[idx - 1];
 
             if (curr.lastLapTimestampMS === 0) {
                 return "";
@@ -252,6 +314,9 @@ export default {
             }
 
             return this.msToTime(gap);
+        },
+        timeSince(date) {
+            return moment(date).format("LTS");
         }
     }
 }
@@ -288,22 +353,23 @@ th {
     background-color: #1b2838;
 }
 
-td, th {
+td,
+th {
     padding: 5px;
 }
 
 tr:nth-child(odd) {
-  background-color: #1f2936;
+    background-color: #1f2936;
 }
 
 #chat .message div {
     display: inline;
     margin-right: 10px;
-} 
+}
 
 #chat .message {
     margin-bottom: 5px;
-} 
+}
 
 .message .ts {
     color: #304363;
@@ -311,6 +377,14 @@ tr:nth-child(odd) {
 
 .message .name {
     font-weight: bold;
+}
+
+.message.damage {
+    color: #888888;
+}
+
+.message.session {
+    color: #803c3c;
 }
 </style>